+++ common/src/main/java/smartin/miapi/mixin/projectile/AbstractArrowAccessor.java
@@ -1,5 +1,6 @@
 package smartin.miapi.mixin.projectile;
 
+import net.minecraft.network.syncher.EntityDataAccessor;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.world.entity.projectile.AbstractArrow;
 import org.spongepowered.asm.mixin.Mixin;
@@ -8,6 +9,11 @@ import org.spongepowered.asm.mixin.gen.Invoker;
 
 @Mixin(AbstractArrow.class)
 public interface AbstractArrowAccessor {
+    @Accessor("PIERCE_LEVEL")
+    static EntityDataAccessor<Byte> getPerceLevelDataPublic() {
+        throw new UnsupportedOperationException();
+    }
+
     @Invoker
     void callSetPierceLevel(byte pierceLevel);