+++ common/src/main/java/smartin/miapi/modules/properties/render/colorproviders/ColorProvider.java
@@ -8,10 +8,8 @@ import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.core.Holder;
 import net.minecraft.core.component.DataComponents;
-import net.minecraft.world.item.ArmorItem;
-import net.minecraft.world.item.ArmorMaterial;
-import net.minecraft.world.item.ItemDisplayContext;
-import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.item.*;
 import net.minecraft.world.item.armortrim.ArmorTrim;
 import org.jetbrains.annotations.Nullable;
 import smartin.miapi.Miapi;
@@ -63,59 +61,57 @@ public interface ColorProvider {
 
     ColorProvider getInstance(ItemStack stack, ModuleInstance instance, TrimRenderer.TrimMode trimMode);
 
-    static boolean isUseTrim(TrimRenderer.TrimMode mode) {
+    static boolean shouldHaveTrim(TrimRenderer.TrimMode mode) {
         return !mode.equals(TrimRenderer.TrimMode.NONE) && !MiapiConfig.getClientConfig().other.disableFastTrim;
     }
 
     @Nullable
-    static SpriteOverlayer getTrimmControllor(Material material, TrimRenderer.TrimMode mode, ItemStack stack, SpriteColorer spriteColorer) {
+    static SpriteOverlayer getTrimController(Material material, TrimRenderer.TrimMode mode, ItemStack stack, SpriteColorer spriteColorer, ItemDisplayContext displayMode) {
         ArmorTrim armorTrim = stack.getComponents().get(DataComponents.TRIM);
         if (armorTrim != null) {
-            Holder<ArmorMaterial> armorMaterial = (stack.getItem() instanceof ArmorItem armorItem) ? armorItem.getMaterial() : null;
-            if (armorMaterial != null) {
-                TextureAtlasSprite trimSprite = switch (mode) {
-                    case ITEM -> {
-                        if(stack.getItem() instanceof ArmorItem armorItem){
-                            if(armorItem.getType() == ArmorItem.Type.HELMET){
-                                yield TrimRenderer.atlas.getSprite(Miapi.id("minecraft:trims/items/helmet_trim_" +armorTrim.material().value().assetName()));
-                            }
-                            if(armorItem.getType() == ArmorItem.Type.CHESTPLATE){
-                                yield TrimRenderer.atlas.getSprite(Miapi.id("minecraft:trims/items/chestplate_trim_" +armorTrim.material().value().assetName()));
-                            }
-                            if(armorItem.getType() == ArmorItem.Type.LEGGINGS){
-                                yield TrimRenderer.atlas.getSprite(Miapi.id("minecraft:trims/items/leggings_trim_" +armorTrim.material().value().assetName()));
-                            }
-                            if(armorItem.getType() == ArmorItem.Type.BOOTS){
-                                yield TrimRenderer.atlas.getSprite(Miapi.id("minecraft:trims/items/boots_trim_" +armorTrim.material().value().assetName()));
-                            }
+            Holder<ArmorMaterial> armorMaterial = (stack.getItem() instanceof ArmorItem armorItem) ? armorItem.getMaterial() : ArmorMaterials.IRON;
+            TextureAtlasSprite trimSprite = switch (mode) {
+                case ITEM -> {
+                    if (stack.getItem() instanceof ArmorItem armorItem) {
+                        if (armorItem.getType() == ArmorItem.Type.HELMET) {
+                            yield TrimRenderer.atlas.getSprite(Miapi.id("minecraft:trims/items/helmet_trim_" + armorTrim.material().value().assetName()));
                         }
-                        yield null;
-                    }
-                    case TrimRenderer.TrimMode.ARMOR_LAYER_ONE ->
-                            TrimRenderer.armorTrimsAtlas.getSprite(armorTrim.outerTexture(armorMaterial));
-                    case TrimRenderer.TrimMode.ARMOR_LAYER_TWO ->
-                            TrimRenderer.armorTrimsAtlas.getSprite(armorTrim.innerTexture(armorMaterial));
-                    default -> {
-                        if(stack.getItem() instanceof ArmorItem armorItem){
-                            if(armorItem.getType() == ArmorItem.Type.HELMET){
-                                yield TrimRenderer.atlas.getSprite(Miapi.id("minecraft:trims/items/helmet_trim_" +armorTrim.material().value().assetName()));
-                            }
-                            if(armorItem.getType() == ArmorItem.Type.CHESTPLATE){
-                                yield TrimRenderer.atlas.getSprite(Miapi.id("minecraft:trims/items/chestplate_trim_" +armorTrim.material().value().assetName()));
-                            }
-                            if(armorItem.getType() == ArmorItem.Type.LEGGINGS){
-                                yield TrimRenderer.atlas.getSprite(Miapi.id("minecraft:trims/items/leggings_trim_" +armorTrim.material().value().assetName()));
-                            }
-                            if(armorItem.getType() == ArmorItem.Type.BOOTS){
-                                yield TrimRenderer.atlas.getSprite(Miapi.id("minecraft:trims/items/boots_trim_" +armorTrim.material().value().assetName()));
-                            }
+                        if (armorItem.getType() == ArmorItem.Type.CHESTPLATE) {
+                            yield TrimRenderer.atlas.getSprite(Miapi.id("minecraft:trims/items/chestplate_trim_" + armorTrim.material().value().assetName()));
                         }
-                        yield null;
+                        if (armorItem.getType() == ArmorItem.Type.LEGGINGS) {
+                            yield TrimRenderer.atlas.getSprite(Miapi.id("minecraft:trims/items/leggings_trim_" + armorTrim.material().value().assetName()));
+                        }
+                        if (armorItem.getType() == ArmorItem.Type.BOOTS) {
+                            yield TrimRenderer.atlas.getSprite(Miapi.id("minecraft:trims/items/boots_trim_" + armorTrim.material().value().assetName()));
+                        }
                     }
-                };
-                if (trimSprite != null) {
-                    return new SpriteOverlayer(material, trimSprite, spriteColorer);
+                    yield null;
                 }
+                case TrimRenderer.TrimMode.ARMOR_LAYER_ONE ->
+                        TrimRenderer.armorTrimsAtlas.getSprite(armorTrim.outerTexture(armorMaterial));
+                case TrimRenderer.TrimMode.ARMOR_LAYER_TWO ->
+                        TrimRenderer.armorTrimsAtlas.getSprite(armorTrim.innerTexture(armorMaterial));
+                default -> {
+                    if (stack.getItem() instanceof Equipable armorItem && displayMode.equals(ItemDisplayContext.GUI)) {
+                        if (armorItem.getEquipmentSlot() == EquipmentSlot.HEAD) {
+                            yield TrimRenderer.atlas.getSprite(Miapi.id("minecraft:trims/items/helmet_trim_" + armorTrim.material().value().assetName()));
+                        }
+                        if (armorItem.getEquipmentSlot() == EquipmentSlot.CHEST) {
+                            yield TrimRenderer.atlas.getSprite(Miapi.id("minecraft:trims/items/chestplate_trim_" + armorTrim.material().value().assetName()));
+                        }
+                        if (armorItem.getEquipmentSlot() == EquipmentSlot.LEGS) {
+                            yield TrimRenderer.atlas.getSprite(Miapi.id("minecraft:trims/items/leggings_trim_" + armorTrim.material().value().assetName()));
+                        }
+                        if (armorItem.getEquipmentSlot() == EquipmentSlot.FEET) {
+                            yield TrimRenderer.atlas.getSprite(Miapi.id("minecraft:trims/items/boots_trim_" + armorTrim.material().value().assetName()));
+                        }
+                    }
+                    yield null;
+                }
+            };
+            if (trimSprite != null) {
+                return new SpriteOverlayer(material, trimSprite, spriteColorer);
             }
         }
         return null;